fix: キャリブレーションプログラムを修正

+++ amg88/calibrate.py
@@ -19,6 +19,7 @@ for d in glob.iglob('csv/*'):
         avg = float(f.read())
     calibrate.append(avg - correct)
 
-os.remove('conf/calibration.txt')
+if os.path.exists('conf/calibration.txt'):
+    os.remove('conf/calibration.txt')
 with open('conf/calibration.txt', 'w') as f:
     f.write(str(sum(calibrate) / len(calibrate)))
